update: トップページをPresentational ComponentとContainer Componentにわけた
