+++ src/components/ResetGameButton.tsx
@@ -1,31 +1,27 @@
-import React, { useState } from 'react'
+import React from 'react'
 import clsx from 'clsx'
-import { initializeDatabase } from '../plugins/firebase'
 
 type Props = {
   label?: string
   className?: string
+  onReset: () => void
 }
 
 export const ResetGameButton: React.FC<Props> = ({
   label = 'リセット',
   className,
+  onReset,
 }) => {
-  const [loading, setLoading] = useState(false)
-
   const handleOnClick = async () => {
     const result = confirm('ゲームをリセットしますか？')
     if (result) {
-      setLoading(true)
-      await initializeDatabase()
-      setLoading(false)
+      onReset()
     }
   }
 
   return (
     <button
       className={clsx('btn bg-gray-800 text-white', className)}
-      disabled={loading}
       onClick={handleOnClick}
     >
       {label}
